20190131 NSY 예약목록,차트 업로드

+++ FINAL/src/main/java/com/spring/bookmanage/reservation/NSYcontroller/NSYReservationController.java
@@ -25,12 +25,14 @@ public class NSYReservationController {
 	@Autowired
 	private AES256 aes;
 	
+	// 예약목록 페이지 요청하기
 	@RequestMapping(value="reservationList.ana",method= {RequestMethod.GET})
 	public String reservationList(HttpServletRequest req, HttpServletResponse res) {
 		
 		return "reservation/reservationList.tiles1";
-	}
+	}//end of String reservationList()
 	
+	// 예역목록 리스트 가져오기 
 	@RequestMapping(value="getReservationList.ana",method= {RequestMethod.GET})
 	@ResponseBody
 	public List<HashMap<String, Object>> getReservationList (HttpServletRequest req, HttpServletResponse res) {
@@ -41,23 +43,24 @@ public class NSYReservationController {
 		String searchWord = req.getParameter("searchWord");
 		String currentShowPageNo = req.getParameter("currentShowPageNo");
 		String sizePerPage = req.getParameter("sizePerPage");
-		String sort = req.getParameter("sort");
 		
+		// 현재 보여지는 페이지번호 지정하기!
 		if(currentShowPageNo==null || "".equals(currentShowPageNo)) {
 			currentShowPageNo= "1";
-		}
+		}//end of if문
 		
 		// 공식 !!!
 		int rno1 = Integer.parseInt(currentShowPageNo) * Integer.parseInt(sizePerPage) - (Integer.parseInt(sizePerPage)-1); // 공식 !!!
 		int rno2 = Integer.parseInt(currentShowPageNo) * Integer.parseInt(sizePerPage); 
 		
+		// 파라미터값 생성하기
 		HashMap<String,String> paraMap = new HashMap<String,String>();
 		
 		paraMap.put("rno1", String.valueOf(rno1));
 		paraMap.put("rno2", String.valueOf(rno2));
 		paraMap.put("colname", colname);
 		paraMap.put("searchWord", searchWord);
-		paraMap.put("sort", sort);
+		
 		
 		List<HashMap<String, Object>> reservationList = service.getReservationList(paraMap);
 		
@@ -92,12 +95,13 @@ public class NSYReservationController {
 			
 			resultMapList.add(resultMap);
 			
-		}
+		}// end of for문
 		
 		return resultMapList;
 		
-	}
+	}// end of List<HashMap<String, Object>> getReservationList()
 	
+	// 예약목록 페이지바 만들기 
 	@RequestMapping(value="getMakeBarPage_ReservationList.ana", method= {RequestMethod.GET})
 	@ResponseBody
 	public HashMap<String, Integer> getMakeBarPage_ReservationList(HttpServletRequest req, HttpServletResponse res){
@@ -111,64 +115,86 @@ public class NSYReservationController {
 		
 		if(currentShowPageNo==null || "".equals(currentShowPageNo)) {
 			currentShowPageNo= "1";
-		}
+		}// end of if문
 		
 		HashMap<String,String > paraMap = new HashMap<String,String>();
 		paraMap.put("sizePerPage", sizePerPage);
 		paraMap.put("colname", colname);
 		paraMap.put("searchWord", searchWord);
 		
-		// ==== 도서관 전체 갯수를 구해오기 ====
+		// 예약된 도서 전체 갯수 구해오기 
 		int totalCount = service.getReservationTotalList(paraMap);
 		
+		// 공식!
 		int totalPage = (int)Math.ceil((double)totalCount/Integer.parseInt(sizePerPage));
 		
 		returnMap.put("totalPage", totalPage);
 		
 		return returnMap;
-	}
+	}// end of HashMap<String, Integer> getMakeBarPage_ReservationList
 	
+	// 예약목록에서 예약되어진 도서를 대여도서로 변경하기
 	@RequestMapping(value="rental.ana", method= {RequestMethod.POST})
 	public String rental(HttpServletRequest req, HttpServletResponse res) {
 		
+		// 목록에서 체크되어진 값을 불러와 선언한다.( 체크항목이 여러개일수있기 때문에 배열로 값을 선언한다)
 		String[] rentalValue = req.getParameterValues("rentalValue");
-		//System.out.println("체크박스값 확인하기 : "+ rentalValue[0]);
 		
-		for(int i=0; i<rentalValue.length; i++) {
-			//System.out.println("체크박스값 확인하기 : "+ rentalValue[i]);
-			String[] value = rentalValue[i].split(",");
-			String bookid = value[0];
-			String memberid = value[1];
-			
-			//System.out.println("value값:"+value);
-			System.out.println("value1값:"+bookid);
-			System.out.println("value2값:"+memberid);
-			
-			HashMap<String, String> paraMap = new HashMap<String, String>();
-			paraMap.put("bookid", bookid);
-			paraMap.put("memberid", memberid);
-			
-			System.out.println("value1값 2번째:"+paraMap.get("bookid"));
-			System.out.println("value2값 2번쨰:"+paraMap.get("memberid"));
-			
-			int result = service.reservation_rental(paraMap);
-			
-			if(result ==1) {
-				service.changBookStatus(paraMap);
-				
-				service.deleteReservation(paraMap);
-				
-				String msg ="대여가 완료되었습니다.";
-				String loc ="";
-			}
-			else {
-				String msg ="대여 실패. 관리자에게 문의하세요";
-				String loc ="";
-			}//end of if_else
-			
-		}// end of for문
+		// 메세지, 위치 선언해주기
+		String msg ="";
+		String loc ="";
 		
-		return "reservation/reservationList.tiles1";
+		// 체크된어진 값이 있다면
+		if(rentalValue != null) {
+			
+			for(int i=0; i<rentalValue.length; i++) {
+				
+				// 배열로 저장된 rentalValue을 ','구분하여  i++ 순서로 선언하기
+				String[] value = rentalValue[i].split(",");
+				
+				// rentalValue구분되어진 value값의 순서를 맞춰 bookid와 memberid로 선언
+				String bookid = value[0];
+				String memberid = value[1];
+				
+				// 파라미터값 생성하기!
+				HashMap<String, String> paraMap = new HashMap<String, String>();
+				paraMap.put("bookid", bookid);
+				paraMap.put("memberid", memberid);
+				
+				// 예약목록의 도서를 대여목록에 추가하기
+				int result = service.reservation_rental(paraMap);
+				
+				// 결과가 성공이라면 
+				if(result ==1) {
+					// book table의 status값 변경 ( =>대여값으로)
+					service.changBookStatus(paraMap);
+					
+					// 예약모록의 도서가 대여가 되어지면 예약목록에서 삭제
+					service.deleteReservation(paraMap);
+					
+					msg ="대여가 완료되었습니다.";
+					loc ="reservationList.ana";
+				}
+				// 결과가 실패라면
+				else {
+					msg ="대여가 완료되지 않았습니다. 관리자에게 문의하세요";
+					loc ="reservationList.ana";
+				}//end of if_else
+				
+			}// end of for문
+			
+		}// if
+		// 체크되어진 값이 없다면
+		else {
+			msg="체크된항목이 없습니다.";
+			loc="reservationList.ana";
+			
+		}// end of if_else
+		
+		req.setAttribute("msg", msg);
+		req.setAttribute("loc", loc);
+		
+		return "msg";
 	}// end of String rental()
 	
 }// end of class NSYReservationController
